requirements.txt
asgiref==3.8.1
awscli==1.38.16
awsebcli==3.21.0
botocore==1.35.99
cement==2.10.14
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.16
environ==1.0
et_xmlfile==2.0.0
idna==3.10
jmespath==1.0.1
openpyxl==3.1.5
pathspec==0.10.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
PyJWT==2.10.1
pypiwin32==223
python-dateutil==2.9.0.post0
pywin32==310
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.11.4
semantic-version==2.10.0
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
